Move serializer and tokenizer into its own file
